backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
timm==1.0.3
groq==0.9.0
langgraph==0.1.19
langchain==0.2.6
langchain-groq==0.1.6
chromadb==0.5.3
sentence-transformers==3.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
cloudinary==1.40.0
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
numpy==1.26.4
opencv-python==4.10.0.84
pymupdf==1.24.1
torchxrayvision